Bug fix: CrashSender queued reports sending via SMTP sends only the first report
The problem was that when trying to send queued crash reports (chosen to be sent later by "Close the program and send report later") via SMTP only the first report was sent and all after that failed.
By reading the batch log file and the source code and by debugging I deduced that CrashSender tried to send all the previous crash reports too when sending n-th report.